requirements.txt
# Core packages
pydantic>=2.11.2
pydantic-settings>=2.0.3

# FAISS for vector similarity
faiss-cpu>=1.7.4

# PDF handling
pdfplumber>=0.10.2

# Gemini (Google Generative AI)
google-generativeai>=0.4.1

# NumPy for embedding processing
numpy>=1.24

google-adk==1.3.0

adktools==0.1.2

db-dtypes==1.4.3

tabulate==0.9.0

uvicorn